requirements.txt
streamlit==1.32.2
numpy==1.23.5
mediapipe==0.10.8
matplotlib==3.8.2
Pillow==10.0.0
